// src/types/auth.types.ts
export interface LoginPayload {
  username: string
  password: string
  expiresInMins?: number
}

export interface User {
  id: number
  username: string
  email: string
  firstName: string
  lastName: string
  gender: string
  image: string
}

export interface AuthResponse extends User {
  accessToken: string
  refreshToken: string
}